+++ agent_workbench/package/__init__.py
@@ -8,6 +8,7 @@
 from __future__ import annotations
 
 from agent_workbench.package.exceptions import PackageValidationError
+from agent_workbench.package.integration import PackageIntegration
 from agent_workbench.package.loader import PackageLoader
 from agent_workbench.package.manifest import PackageManifest
 from agent_workbench.package.package_info import PackageInfo
@@ -16,6 +17,7 @@ from agent_workbench.package.registry import PackageRegistry
 
 __all__ = [
     "PackageInfo",
+    "PackageIntegration",
     "PackageLoader",
     "PackageManifest",
     "PackageRegistry",